EV-110 Fix pytest problem (run without calling test file), do minor test refactoring (folder with seperate test files)
