+++ Dockerfile
@@ -12,6 +12,4 @@ RUN pip install --no-cache-dir --upgrade -r ./requirements.txt
 
 COPY ./src /code/src
 
-COPY ./testing.py /code/testing.py
-
 CMD [ "uvicorn", "src.server.app:app", "--host", "0.0.0.0", "--port", "80" ]
